Rename `mesh_pipeline_layout.rs` to `mesh_view_layout.rs`
